[FLOW-22] refactor: styles split
