+++ src/app/page.tsx
@@ -1,5 +1,6 @@
-import { HomePage } from '@/page';
+import { ErdProjectPage } from '@/page';
 
 export default function Home() {
-  return <HomePage />;
+  return <ErdProjectPage />;
+  // return <HomePage />;
 }
