+++ src/components/column-edit/FkColumns.tsx
@@ -1,4 +1,4 @@
-import styled from '@emotion/styled';
+import { styles } from './FkColumns.styles';
 
 import { ERDColumn, ERDTable } from '@/features/erd-project';
 import { useERDProjectStore } from '@/providers';
@@ -65,32 +65,3 @@ export function FkColumns({ table, fkColumns }: FkColumnsProps) {
     </styles.columnRow>
   ));
 }
-
-const styles = {
-  columnRow: styled.div`
-    display: flex;
-    align-items: center;
-    padding-bottom: 8px;
-  `,
-
-  input: styled.input`
-    background: #444;
-    border: none;
-    padding: 8px;
-    color: #ededed;
-    margin-right: 8px;
-    width: 150px;
-
-    &::placeholder {
-      color: #888;
-    }
-  `,
-
-  notNullText: styled.div`
-    white-space: nowrap;
-    margin-left: auto;
-    color: #ededed;
-    font-size: 14px;
-    margin-right: 5px;
-  `,
-};
