+++ src/components/column-edit/PkFkColumns.tsx
@@ -1,4 +1,4 @@
-import styled from '@emotion/styled';
+import { styles } from './PkFkColumns.styles';
 
 import { ERDColumn, ERDTable } from '@/features/erd-project';
 import { useERDProjectStore } from '@/providers';
@@ -58,24 +58,3 @@ export function PkFkColumns({ table, pkfkColumns }: PkFkColumnsProps) {
     </styles.columnRow>
   ));
 }
-
-const styles = {
-  columnRow: styled.div`
-    display: flex;
-    align-items: center;
-    padding-bottom: 8px;
-  `,
-
-  input: styled.input`
-    background: #444;
-    border: none;
-    padding: 8px;
-    color: #ededed;
-    margin-right: 8px;
-    width: 150px;
-
-    &::placeholder {
-      color: #888;
-    }
-  `,
-};
