+++ src/components/implements-icon/icon.styles.ts
@@ -1,0 +1,21 @@
+import styled from '@emotion/styled';
+
+export const styles = {
+  container: styled.div`
+    user-select: none;
+    cursor: pointer;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  `,
+
+  pathStyle: styled.path`
+    &:hover {
+      fill: #ddd;
+    }
+  `,
+
+  menuIcon: styled.div`
+    cursor: pointer;
+  `,
+};
