--- src/components/index.ts
@@ -1,4 +1,0 @@
-export * from './TableInformation';
-export * from './RelationshipInformation';
-export * from './ErdDrawTools';
-export * from './Relationship';
