+++ src/components/project-page/ErdDrawTools.styles.ts
@@ -1,0 +1,23 @@
+import styled from '@emotion/styled';
+
+export const styles = {
+  container: styled.div`
+    display: flex;
+    height: 111px;
+    padding: 16px;
+    justify-content: center;
+    align-items: center;
+    gap: 16px;
+    flex: 1 0 0;
+    border-radius: 16px;
+    border: 0.5px solid #606060;
+    background: #222;
+    z-index: 1;
+  `,
+
+  notationChange: styled.div`
+    cursor: pointer;
+    font-size: 16px;
+    color: #ededed;
+  `,
+};
