+++ src/components/project-page/ErdDrawTools.tsx
@@ -1,14 +1,13 @@
 'use client';
 
-import styled from '@emotion/styled';
-
+import { styles } from './ErdDrawTools.styles';
 import {
   ArrowIcon,
   MemoIcon,
   PointerIcon,
   RelationshipIcons,
   TableIcon,
-} from './implements-icon';
+} from '../implements-icon';
 
 import { useDrawToolsStore } from '@/features/draw-tools';
 
@@ -50,25 +49,3 @@ export function ErdDrawTools() {
     </styles.container>
   );
 }
-
-const styles = {
-  container: styled.div`
-    display: flex;
-    height: 111px;
-    padding: 16px;
-    justify-content: center;
-    align-items: center;
-    gap: 16px;
-    flex: 1 0 0;
-    border-radius: 16px;
-    border: 0.5px solid #606060;
-    background: #222;
-    z-index: 1;
-  `,
-
-  notationChange: styled.div`
-    cursor: pointer;
-    font-size: 16px;
-    color: #ededed;
-  `,
-};
