+++ src/components/project-page/Relationship.styles.ts
@@ -1,0 +1,13 @@
+import styled from '@emotion/styled';
+
+export const styles = {
+  display: styled.div`
+    width: 100%;
+    height: 100%;
+    position: absolute;
+  `,
+  wrapper: styled.svg`
+    width: 100%;
+    height: 100%;
+  `,
+};
