+++ src/components/project-page/RelationshipInformation.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import { styles } from './RelationshipInformation.styles';
+
+export function RelationshipInformation() {
+  return <styles.container />;
+}
