+++ src/components/project-page/RelationshipMenu.tsx
@@ -1,4 +1,4 @@
-import styled from '@emotion/styled';
+import { styles } from './RelationshipMenu.styles';
 
 import { useERDProjectStore } from '@/providers';
 import { useMappingContext } from '@/providers/MappingProvider';
@@ -71,28 +71,3 @@ export function RelationshipMenu() {
     </styles.menu>
   );
 }
-
-const styles = {
-  menu: styled.div<{ $pos: { left: number; top: number } }>`
-    position: absolute;
-    left: ${({ $pos }) => `${$pos.left}px`};
-    top: ${({ $pos }) => `${$pos.top}px`};
-    background-color: #333;
-    border: 1px solid #444;
-    border-radius: 4px;
-    box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.3);
-    z-index: 1;
-  `,
-
-  menuItem: styled.div`
-    padding: 8px 16px;
-    font-size: 12px;
-    color: #fff;
-    white-space: nowrap;
-    cursor: pointer;
-
-    &:hover {
-      background-color: #444;
-    }
-  `,
-};
