+++ src/components/project-page/TableInformation.styles.ts
@@ -1,12 +1,6 @@
-'use client';
-
 import styled from '@emotion/styled';
 
-export function TableInformation() {
-  return <styles.container />;
-}
-
-const styles = {
+export const styles = {
   container: styled.div`
     display: flex;
     width: 200px;
