+++ src/components/project-page/TableInformation.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import { styles } from './TableInformation.styles';
+
+export function TableInformation() {
+  return <styles.container />;
+}
