+++ src/components/table/Table.tsx
@@ -1,9 +1,9 @@
 'use client';
 
-import styled from '@emotion/styled';
 import { useEffect, useLayoutEffect, useRef, useState } from 'react';
 
 import { Columns } from './Columns';
+import { styles } from './Table.styles';
 import { TableMenu } from './TableMenu';
 
 import { ColumnEditMenu } from '@/components/column-edit';
@@ -155,51 +155,3 @@ export function Table({ table, onClick, onPositionChange }: TableProps) {
     </styles.displayWrapper>
   );
 }
-
-const styles = {
-  displayWrapper: styled.div<{ $pos: Position }>`
-    position: absolute;
-    display: flex;
-    flex-direction: column;
-    left: ${({ $pos }) => `${$pos.left}px`};
-    top: ${({ $pos }) => `${$pos.top - 20}px`};
-  `,
-
-  titleMenuWrapper: styled.div`
-    display: flex;
-    flex-direction: row;
-  `,
-
-  container: styled.div`
-    display: flex;
-    flex-direction: column;
-    gap: 5px;
-    min-width: 50px;
-    min-height: 30px;
-    border: 0.5px solid #606060;
-    background: rgba(34, 34, 34, 0.7);
-    padding: 10px;
-    cursor: pointer;
-
-    &:hover {
-      color: #fff;
-    }
-  `,
-
-  tableTitle: styled.div`
-    font-size: 12px;
-    color: #ededed;
-    flex-grow: 1;
-    justify-items: start;
-    margin-right: 10px;
-    width: 100%;
-    left: 0;
-    top: -20px;
-  `,
-
-  contour: styled.div`
-    width: 100%;
-    height: 1px;
-    background-color: #ccc;
-  `,
-};
