+++ src/page/erd-project/index.ts
@@ -1,0 +1,1 @@
+export * from './erd-project.page';
